feat: :sparkles: Ingredients Table Update
Add name, price, ingredientPic column

+++ database/migrations/2022_10_11_110044_create_orders_table.php
@@ -19,7 +19,7 @@ class CreateOrdersTable extends Migration
             $table->string('surname');
             $table->text('address');
             $table->text('comment')->nullable();
-            $table->float('price')->unsigned();
+            $table->double('price')->unsigned();
             $table->timestamps();
         });
     }
